requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
requests==2.31.0
pillow==10.1.0
torch==2.1.1
torchvision==0.16.1
msgpack==1.0.7
git+https://github.com/openai/CLIP.git
python-multipart==0.0.6
